Remove unneeded casts.

+++ lib/features/graphs/presentation/pages/graphs_page.dart
@@ -252,13 +252,15 @@ class _GraphsViewState extends State<GraphsView> {
                   enabled: state.status == BlocStatus.success,
                   icon: FaIcon(
                     state.status == BlocStatus.failure ? FontAwesomeIcons.userSlash : FontAwesomeIcons.solidUser,
-                    color: (_userId != -1 && _userId != null) ? Theme.of(context).colorScheme.primary : Theme.of(context).colorScheme.onSurface,
+                    color: (_userId != -1 && _userId != null)
+                        ? Theme.of(context).colorScheme.primary
+                        : Theme.of(context).colorScheme.onSurface,
                     size: 20,
                   ),
                   tooltip: LocaleKeys.select_user_title.tr(),
                   onSelected: (value) {
                     setState(() {
-                      _userId = value as int;
+                      _userId = value;
                     });
 
                     _graphsBloc.add(
@@ -287,9 +289,13 @@ class _GraphsViewState extends State<GraphsView> {
                                 state as SettingsSuccess;
 
                                 return Text(
-                                  state.appSettings.maskSensitiveInfo ? LocaleKeys.hidden_message.tr() : user.friendlyName ?? '',
+                                  state.appSettings.maskSensitiveInfo
+                                      ? LocaleKeys.hidden_message.tr()
+                                      : user.friendlyName ?? '',
                                   style: TextStyle(
-                                    color: _userId == user.userId! ? Theme.of(context).colorScheme.primary : Theme.of(context).colorScheme.onSurface,
+                                    color: _userId == user.userId!
+                                        ? Theme.of(context).colorScheme.primary
+                                        : Theme.of(context).colorScheme.onSurface,
                                   ),
                                 );
                               },
@@ -360,13 +366,17 @@ class _GraphsViewState extends State<GraphsView> {
                       FaIcon(
                         FontAwesomeIcons.hashtag,
                         size: 20,
-                        color: _yAxis == PlayMetricType.plays ? Theme.of(context).colorScheme.primary : Theme.of(context).colorScheme.onSurface,
+                        color: _yAxis == PlayMetricType.plays
+                            ? Theme.of(context).colorScheme.primary
+                            : Theme.of(context).colorScheme.onSurface,
                       ),
                       const Gap(8),
                       Text(
                         LocaleKeys.play_count_title,
                         style: TextStyle(
-                          color: _yAxis == PlayMetricType.plays ? Theme.of(context).colorScheme.primary : Theme.of(context).colorScheme.onSurface,
+                          color: _yAxis == PlayMetricType.plays
+                              ? Theme.of(context).colorScheme.primary
+                              : Theme.of(context).colorScheme.onSurface,
                         ),
                       ).tr(),
                     ],
@@ -379,13 +389,17 @@ class _GraphsViewState extends State<GraphsView> {
                       FaIcon(
                         FontAwesomeIcons.solidClock,
                         size: 20,
-                        color: _yAxis == PlayMetricType.time ? Theme.of(context).colorScheme.primary : Theme.of(context).colorScheme.onSurface,
+                        color: _yAxis == PlayMetricType.time
+                            ? Theme.of(context).colorScheme.primary
+                            : Theme.of(context).colorScheme.onSurface,
                       ),
                       const Gap(8),
                       Text(
                         LocaleKeys.play_time_title,
                         style: TextStyle(
-                          color: _yAxis == PlayMetricType.time ? Theme.of(context).colorScheme.primary : Theme.of(context).colorScheme.onSurface,
+                          color: _yAxis == PlayMetricType.time
+                              ? Theme.of(context).colorScheme.primary
+                              : Theme.of(context).colorScheme.onSurface,
                         ),
                       ).tr(),
                     ],
@@ -466,7 +480,9 @@ class _GraphsViewState extends State<GraphsView> {
                         child: Text(
                           '7 ${LocaleKeys.days_title.tr()}',
                           style: TextStyle(
-                            color: _timeRange == 7 ? Theme.of(context).colorScheme.primary : Theme.of(context).colorScheme.onSurface,
+                            color: _timeRange == 7
+                                ? Theme.of(context).colorScheme.primary
+                                : Theme.of(context).colorScheme.onSurface,
                           ),
                         ),
                       ),
@@ -475,7 +491,9 @@ class _GraphsViewState extends State<GraphsView> {
                         child: Text(
                           '14 ${LocaleKeys.days_title.tr()}',
                           style: TextStyle(
-                            color: _timeRange == 14 ? Theme.of(context).colorScheme.primary : Theme.of(context).colorScheme.onSurface,
+                            color: _timeRange == 14
+                                ? Theme.of(context).colorScheme.primary
+                                : Theme.of(context).colorScheme.onSurface,
                           ),
                         ),
                       ),
@@ -484,7 +502,9 @@ class _GraphsViewState extends State<GraphsView> {
                         child: Text(
                           '30 ${LocaleKeys.days_title.tr()}',
                           style: TextStyle(
-                            color: _timeRange == 30 ? Theme.of(context).colorScheme.primary : Theme.of(context).colorScheme.onSurface,
+                            color: _timeRange == 30
+                                ? Theme.of(context).colorScheme.primary
+                                : Theme.of(context).colorScheme.onSurface,
                           ),
                         ),
                       ),
@@ -493,7 +513,9 @@ class _GraphsViewState extends State<GraphsView> {
                         child: Text(
                           'Custom',
                           style: TextStyle(
-                            color: ![7, 14, 30].contains(_timeRange) ? Theme.of(context).colorScheme.primary : Theme.of(context).colorScheme.onSurface,
+                            color: ![7, 14, 30].contains(_timeRange)
+                                ? Theme.of(context).colorScheme.primary
+                                : Theme.of(context).colorScheme.onSurface,
                           ),
                         ),
                       ),
